+++ dbtests/common/fixtures.go
@@ -2,7 +2,7 @@ package common
 
 import (
 	"time"
-	"github.com/samonzeweb/godb/nullable"
+	"github.com/samonzeweb/godb/types"
 )
 
 type Book struct {
@@ -29,8 +29,8 @@ func (*Inventory) TableName() string {
 }
 
 type InventoryPart struct {
-	Id       nullable.NullInt64 `db:"id"`
-	Counting nullable.NullInt64 `db:"counting"`
+	Id       types.NullInt64 `db:"id"`
+	Counting types.NullInt64 `db:"counting"`
 }
 
 type BooksWithInventories struct {
